move firebase set up to channel client
